+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking** `getCompensations` now returns a list of lists, with the `spillMatrix` formatted as a matrix.
 
 ### Removed
+- The `xVertices` and `yVertices` arguments to `createPolygonGate` are replaced
+  by `vertices=list(c(), c(), c())`.
 
 ## [Unreleased]
 
@@ -55,7 +57,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix mangling of experiment properties in `updateGate()`.
 - Remove unused `params` argument from `getExperiment()`.
 - Run `styler` on all files with default params
-- Deprecate `createPolygonGate` args `xVertices=c()` and
-  `yVertices=c()` in favor of `vertices=list(c(), c(), c())`
 
 ### Removed